calculate all

+++ database/migrations/2023_11_22_120513_create_decision_matrices_table.php
@@ -14,7 +14,7 @@ return new class extends Migration {
             $table->id();
             $table->unsignedBigInteger('alternative_id');
             $table->unsignedBigInteger('criterion_id');
-            $table->decimal('value', 5);
+            $table->integer('value');
             $table->foreign('alternative_id')->references('id')->on('alternatives')->onDelete('cascade');
             $table->foreign('criterion_id')->references('id')->on('criteria')->onDelete('cascade');
         });
